backend/requirements.txt
# Core
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.28.0

# Phase 1: Semantic re-ranking (sentence-transformers)
sentence-transformers>=2.2.0
numpy>=1.24.0

# Optional: LLM agent (Gemini)
# google-genai>=0.3.0
